% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_reassign.R
\name{catch_reassign}
\alias{catch_reassign}
\title{creates an SQL script that will reassign a list of catches to a new reference catch}
\usage{
catch_reassign(catch_ids = NULL, aphia_id = NULL, quiet = FALSE)
}
\arguments{
\item{catch_ids}{A lsit of catch ids to update}

\item{aphia_id}{The new aphia_id for the given catch_ids}

\item{quiet}{Supress console messages (default is FALSE).}
}
\value{
an SQL statement that will need to be executed.
}
\description{
This is a function that creates an SQL statments that can updates a list of catches.
It is not strickly used by the package, but is a util for end-users.
}
