% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.calc_nautical_dist}
\alias{.calc_nautical_dist}
\title{Calculate nautical distance between two geographic points
#' This function calculates the great-circle distance between two points on the Earth's surface
#' given their latitude and longitude using the Haversine formula.}
\usage{
.calc_nautical_dist(p1_lat, p1_lon, p2_lat, p2_lon)
}
\arguments{
\item{p1_lat}{Latitude of the first point in decimal degrees}

\item{p1_lon}{Longitude of the first point in decimal degrees}

\item{p2_lat}{Latitude of the second point in decimal degrees}

\item{p2_lon}{Longitude of the second point in decimal degrees}
}
\description{
Calculate nautical distance between two geographic points
#' This function calculates the great-circle distance between two points on the Earth's surface
#' given their latitude and longitude using the Haversine formula.
}
