% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_from_mission.R
\name{event_from_mission}
\alias{event_from_mission}
\title{Create an OBISEvent from an Andes mission}
\usage{
event_from_mission(mission, parent = NULL, quiet = FALSE)
}
\arguments{
\item{mission}{A dataframe (single row) containing a mission set from an Andes database}

\item{parent}{The parent event of class OBISEvent (default is NULL). The mission OBISevent is usually usually the top-level event and has no parent.}

\item{quiet}{Supress console messages (default is FALSE).}
}
\value{
an event of class OBISEvent representing a mission.
}
\description{
This function creates an OBISEvent object representing a mission from the Andes database.
It maps certain fields to a corresponding DwC event fields, and addes hard-coded values where necessary.
}
\seealso{
[get_andes_mission()] for getting a list of missions  (usualy only a single active mission) from the Andes DB.

Other OBISEvent creation functions:
\code{\link[=event_from_fishing_set]{event_from_fishing_set()}}
}
\concept{OBISEvent creation functions}
