% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biodiv_data.R
\name{get_biodiv_data}
\alias{get_biodiv_data}
\title{Gets biodiversity data from ANDES database}
\usage{
get_biodiv_data(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
a dataframe containing andes biodiveristy data
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct a Biodiversity table.
TODO: The mission filter is not yet implemented everywhere, but should be added in the future. It currently uses the active mission
The current ANDES active mission will determine for which mission the data are returned.
}
\seealso{
[andes_db_connect()] for getting a connection object to the ANDES database.

Other Andesdb query functions:
\code{\link[=get_andes_catch]{get_andes_catch()}},
\code{\link[=get_andes_mission]{get_andes_mission()}},
\code{\link[=get_andes_set]{get_andes_set()}}
}
\concept{Andesdb query functions}
