% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biodiv_data.R
\name{get_biodiv_data_db}
\alias{get_biodiv_data_db}
\title{Gets get_biodiv_data_db (raw database results)}
\usage{
get_biodiv_data_db(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing the biodiversity data set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the Biodiversity table.
The current ANDES active mission will determine for which mission the data are returned.
}
\details{
This function is intended for internal use and returns raw results from the database.
It is not meant for direct use in analysis or reporting. Users should use `get_biodiv_data`
}
\seealso{
[get_capture_mollusque()] for the formatted results
}
