% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biodiv_specimen_data.R
\name{get_biodiv_specimen_data}
\alias{get_biodiv_specimen_data}
\title{Gets all catch data from ANDES database, same as get-biodiv_data but does not filter by class and includes specimen count}
\usage{
get_biodiv_specimen_data(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
a dataframe containing andes biodiveristy data
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct a Biodiversity table.
TODO: The mission filter is not yet implemented everywhere, but should be added in the future. It currently uses the active mission
The current ANDES active mission will determine for which mission the data are returned.
}
\seealso{
[andes_db_connect()] for getting a connection object to the ANDES database.

Other Andesdb query functions:
\code{\link[=get_andes_catch]{get_andes_catch()}},
\code{\link[=get_andes_mission]{get_andes_mission()}},
\code{\link[=get_andes_set]{get_andes_set()}},
\code{\link[=get_biodiv_data]{get_biodiv_data()}}
}
\concept{Andesdb query functions}
