% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OBIS_archive.R
\name{get_OBIS_archive}
\alias{get_OBIS_archive}
\title{Extract OBIS archive from ANDES database}
\usage{
get_OBIS_archive(andes_db_connection, ...)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing the OBIS data. Use `$occurrence` and `$event` to access the respective tables.
}
\description{
This function extracts data from an ANDES database connection and formats it into OBIS-compliant occurrence and event tables.
}
