% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join.R
\name{left_join_preserve_order}
\alias{left_join_preserve_order}
\title{a merge that preserves row and column order}
\usage{
left_join_preserve_order(x, y, ...)
}
\arguments{
\item{x}{the "left" dataframe (all.x=TRUE)}

\item{y}{the "right" dataframe}

\item{...}{Additional arguments passed on to methgods}
}
\description{
shamelessly stolen from https://stackoverflow.com/questions/17878048/merge-two-data-frames-while-keeping-the-original-row-order
}
