% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linestring_wkt.R
\name{linestring_wkt}
\alias{linestring_wkt}
\title{Create a WKT linestring from two points
This function creates a WKT (Well-Known Text) representation of a linestring
connecting two points. Each point is represented as a list containing 'lon' and 'lat' fields,
and optionally a 'z' field for elevation (a negative number representing depth).}
\usage{
linestring_wkt(p1, p2, fmt = 3)
}
\arguments{
\item{p1}{A list with 'lon' and 'lat' (and optionally 'z') fields for the first point}

\item{p2}{A list with 'lon' and 'lat' (and optionally 'z') fields for the second point

defing this here instead of using the package sf to avoid dependency
If we ever import sf or wellknown, we can replace this function}
}
\description{
Create a WKT linestring from two points
This function creates a WKT (Well-Known Text) representation of a linestring
connecting two points. Each point is represented as a list containing 'lon' and 'lat' fields,
and optionally a 'z' field for elevation (a negative number representing depth).
}
