% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_set_wkt.R
\name{make_set_wkt}
\alias{make_set_wkt}
\title{Create a WKT linestring for an Andes set}
\usage{
make_set_wkt(set)
}
\arguments{
\item{set}{A dataframe (single row) containing a fishing set from ANDES database}
}
\description{
This function creates a WKT (Well-Known Text) representation of a linestring
connecting the start and end coordinates of a fishing set from the Andes database.
}
