% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obis_time_str}
\alias{obis_time_str}
\title{Convert a datetime to an OBIS-compliant time string based on precision}
\usage{
obis_time_str(dt, precision, tz = NULL)
}
\arguments{
\item{dt}{A POSIXct datetime object.}

\item{precision}{An integer indicating the precision level (1-7).}

\item{tz}{An optional timezone string.}
}
\value{
A string formatted according to OBIS standards.
}
\description{
Convert a datetime to an OBIS-compliant time string based on precision
}
