% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occurrence_from_catch.R
\name{occurrence_from_catch}
\alias{occurrence_from_catch}
\title{Create an OBISOccurence from an Andes catch}
\usage{
occurrence_from_catch(catch, event = NULL, quiet = FALSE)
}
\arguments{
\item{catch}{A dataframe (single row) containing a catch data from an Andes database}

\item{event}{The parent event of class OBISEvent (default is NULL). The mission OBISevent is usually usually the top-level event and has no parent.}

\item{quiet}{Supress console messages (default is FALSE).}
}
\value{
an occurrence of class OBISOccurence representing the catch.
}
\description{
This function creates an OBISOccurence object representing a catch from the Andes database.
It maps certain fields to a corresponding DwC event fields, and addes hard-coded values where necessary.
}
\concept{{OBISOccurrence creation functions}}
