% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{row_union}
\alias{row_union}
\title{rbind, but allow different columns names
Make a union of the rows filling missing columns with NA}
\usage{
row_union(df1, df2)
}
\description{
https://stackoverflow.com/questions/3402371/combine-two-data-frames-by-rows-rbind-when-they-have-different-sets-of-columns
}
