% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_from_fishing_set.R
\name{event_from_fishing_set}
\alias{event_from_fishing_set}
\title{Create an OBISEvent from an Andes fishing set}
\usage{
event_from_fishing_set(set, parent_event = NULL, quiet = FALSE)
}
\arguments{
\item{set}{A dataframe (single row) containing a fishing set from ANDES database}

\item{parent_event}{The parent event of class OBISEvent (default is NULL).}

\item{quiet}{Supress console messages (default is FALSE).}
}
\value{
an event of class OBISEvent representing a fishing set
}
\description{
This function creates an OBISEvent object representing a fishing set from the ANDES database.
It maps certain fields to a corresponding DwC event fields, and addes hard-coded values where necessary.
}
\seealso{
[get_andes_set()] for getting a list of sets from the Andes DB.

Other {OBISEvent creation functions}: 
\code{\link{event_from_mission}()}
}
\concept{{OBISEvent creation functions}}
