% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_andes_catch.R
\name{get_andes_catch}
\alias{get_andes_catch}
\title{Get a catch dataframe}
\usage{
get_andes_catch(andes_db_connection, mission_id = NULL)
}
\arguments{
\item{andes_db_connection}{A connection object to the ANDES database.}

\item{mission_id}{A string or integer representing the mission ID to filter the catches (default is NULL, which retrieves all catches).}
}
\value{
a dataframe containing andes catch rows
}
\description{
This function queries the andes database and returns a dataframe containing the catch data.
TODO: The mission filter is not yet implemented everywhere, but should be added in the future. It currently uses the active mission
}
\seealso{
[andes_db_connect()] for getting a connection object to the ANDES database.

Other {Andesdb query functions}: 
\code{\link{get_andes_mission}()},
\code{\link{get_andes_set}()},
\code{\link{get_biodiv_data}()}
}
\concept{{Andesdb query functions}}
