% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biodiv_data.R
\name{get_biodiv_data}
\alias{get_biodiv_data}
\title{Gets get_biodiv_data (raw database results)}
\usage{
get_biodiv_data(andes_db_connection)
}
\arguments{
\item{andes_db_connection}{a connection object to the ANDES database.}
}
\value{
A dataframe containing the biodiversity data set data.
}
\description{
This function executes a SQL query to retrieve the needed andes data to construct the Biodiversity table.
The current ANDES active mission will determine for which mission the data are returned.
}
\seealso{
[get_biodiv_data_db()] for the formatted results
}
